+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.9.24
+
+* `Ktor`:
+    * `Common`:
+        * New extension fun `MPPFile#input`
+
 ## 0.9.23
 
 * `Repos`: